requirements.txt
paho-mqtt>=1.6.1
pydantic>=2.0.0
click>=8.0.0
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 